Better logging and small fix

+++ src/database_struct/server_activity_struct.rs
@@ -1,3 +1,4 @@
+#[derive(sqlx::FromRow)]
 pub struct ServerActivity {
     pub anime_id: Option<String>,
     pub timestamp: Option<String>,
@@ -5,9 +6,10 @@ pub struct ServerActivity {
     pub webhook: Option<String>,
     pub episode: Option<String>,
     pub name: Option<String>,
-    pub delays: Option<u32>,
+    pub delays: Option<i64>,
 }
 
+#[derive(sqlx::FromRow)]
 pub struct ServerActivityFull {
     pub anime_id: i32,
     pub timestamp: i64,
